Centralised SESSIONID management for urls
Added a method in APIConfigService to construct urls with or without added
SESSIONID. It's a simple comment/uncomment mechanism for now which lets us
decide in just one place whether we want or don't want to have it.

I replaced all current urls that had SESSIONID added to use this method, all
future ones should also use it.